Renaming project to canvas_course_admin_tools
