+++ README.md
@@ -1,3 +1,3 @@
-# Canvas Admin Tools
+# Canvas Course Admin Tools
 
-This project provides a collection of tools used to administer Harvard TLT's instance of Canvas.
+This project provides a collection of tools used to administer Harvard TLT's instance of Canvas at the course level.
