--- canvas_admin_tools/fixtures/external_tools_qa.json
@@ -1,32 +1,0 @@
-[
-{
-    "fields": {
-        "canvas_account_id": "*",
-        "canvas_course_id": "*",
-        "name": "canvas_site_creator",
-        "external_tool_id": 1217
-    },
-    "model": "canvas_admin_tools.externaltool",
-    "pk": 13
-},
-{
-    "fields": {
-        "canvas_account_id": "*",
-        "canvas_course_id": "*",
-        "name": "courses_in_this_account",
-        "external_tool_id": 279
-    },
-    "model": "canvas_admin_tools.externaltool",
-    "pk": 14
-},
-{
-    "fields": {
-        "canvas_account_id": "*",
-        "canvas_course_id": "*",
-        "name": "lti_tools_usage",
-        "external_tool_id": 1521
-    },
-    "model": "canvas_admin_tools.externaltool",
-    "pk": 15
-}
-]
