--- canvas_admin_tools/requirements/demo.txt
@@ -1,4 +1,0 @@
-# stage environment requirements
-
-#includes the base.txt requirements needed in all environments
--r aws.txt
