--- canvas_admin_tools/requirements/production.txt
@@ -1,4 +1,0 @@
-# production environment requirements
-
-#includes the base.txt requirements needed in all environments
--r aws.txt
