--- canvas_admin_tools/requirements/qa.txt
@@ -1,4 +1,0 @@
-# qa environment requirements
-
-#includes the base.txt requirements needed in all environments
--r aws.txt
