--- canvas_admin_tools/requirements/test.txt
@@ -1,4 +1,0 @@
-# test environment requirements
-
-#includes the base.txt requirements needed in all environments
--r aws.txt
