--- canvas_admin_tools/settings/demo.py
@@ -1,4 +1,0 @@
-from .aws import *
-
-# Allow the mailing_list app to allow any user to be added to listserv
-IGNORE_WHITELIST = True
