--- canvas_admin_tools/settings/qa.py
@@ -1,1 +1,0 @@
-from .aws import *
