--- canvas_admin_tools/settings/test.py
@@ -1,1 +1,0 @@
-from .aws import *
