--- canvas_admin_tools/templates/canvas_admin_tools/dashboard_account.html
@@ -1,73 +1,0 @@
-{% extends 'canvas_admin_tools/base.html' %}
-
-{% load static %}
-
-{% block stylesheets %}
-<link href="{% static 'canvas_admin_tools/css/dashboard.css' %}" rel="stylesheet" />
-{% endblock stylesheets %}
-
-{% block content %}
-<header>
-    <div class="row lti-header">
-        <h1>Administration Tasks</h1>
-    </div>
-</header>
-
-<main>
-    {% if has_manage_courses %}
-    <div class="col-md-6">
-        <div class="panel panel-default">
-            <div class="panel-body">
-                <h3>Manage Courses</h3>
-                <div class="list-group">
-                    {% if canvas_site_creator %}
-                    <a href="{{ canvas_site_creator }}" class="list-group-item active" target="_parent">
-                        <h4 class="list-group-item-heading">Canvas Site Creator</h4>
-                        <p class="list-group-item-text">Create Canvas sites for one or multiple courses</p>
-                    </a>
-                    {% endif %}
-                </div>
-            </div>
-        </div>
-    </div>
-    {% endif %}
-    {% if has_manage_account %}
-    <div class="col-md-6">
-        <div class="panel panel-default">
-            <div class="panel-body">
-                <h3>Manage My Account</h3>
-                <div class="list-group">
-                    {% if conclude_courses %}
-                    <a href="{{ conclude_courses }}" class="list-group-item active" target="_blank">
-                        <h4 class="list-group-item-heading">Conclude Courses</h4>
-                        <p class="list-group-item-text">Change conclusion dates for a selected course</p>
-                    </a>
-                    {% endif %}
-                    {% if lti_tools_usage %}
-                    <a href="{{ lti_tools_usage }}" class="list-group-item active" target="_parent">
-                        <h4 class="list-group-item-heading">LTI Tools Usage Report</h4>
-                        <p class="list-group-item-text">See which LTI tools are installed in your account</p>
-                    </a>
-                    {% endif %}
-                    {% if courses_in_this_account %}
-                    <a href="{{ courses_in_this_account }}" class="list-group-item active" target="_parent">
-                        <h4 class="list-group-item-heading">Courses in this Account</h4>
-                        <p class="list-group-item-text">See information about the courses in your account combined in one view</p>
-                    </a>
-                    {% endif %}
-                </div>
-            </div>
-        </div>
-    </div>
-    {% endif %}
-    {% if not has_manage_courses and not has_manage_account %}
-    <div class="col-md-6">
-        <div class="panel panel-default">
-            <div class="panel-body">
-                <p>There are no administration tools installed in this account.</p>
-            </div>
-        </div>
-    </div>
-    {% endif %}
-</main>
-{% endblock content %}
