+++ canvas_course_admin_tools/settings/aws.py
