+++ canvas_course_admin_tools/settings/local.py
