+++ canvas_course_admin_tools/settings/secure.py.example
