+++ canvas_course_admin_tools/templates/canvas_course_admin_tools/base.html
