+++ canvas_course_admin_tools/templates/canvas_course_admin_tools/dashboard_course.html
@@ -1,9 +1,9 @@
-{% extends 'canvas_admin_tools/base.html' %}
+{% extends 'canvas_course_admin_tools/base.html' %}
 
 {% load static %}
 
 {% block stylesheets %}
-<link href="{% static 'canvas_admin_tools/css/dashboard.css' %}" rel="stylesheet" />
+<link href="{% static 'canvas_course_admin_tools/css/dashboard.css' %}" rel="stylesheet" />
 {% endblock stylesheets %}
 
 {% block content %}
