+++ isites_migration/templates/isites_migration/index.html
@@ -1,4 +1,4 @@
-{% extends 'canvas_admin_tools/base.html' %}
+{% extends 'canvas_course_admin_tools/base.html' %}
 
 {% load static %}
 {% load django_helpers %}
