+++ manage.py
@@ -4,9 +4,9 @@ import sys
 
 if __name__ == "__main__":
     if 'test' in sys.argv:
-        os.environ['DJANGO_SETTINGS_MODULE'] = 'canvas_admin_tools.settings.unit_test'
+        os.environ['DJANGO_SETTINGS_MODULE'] = 'canvas_course_admin_tools.settings.test'
     else:
-        os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'canvas_admin_tools.settings.aws')
+        os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'canvas_course_admin_tools.settings.aws')
 
     from django.core.management import execute_from_command_line
