+++ vagrant/venv_bootstrap.sh
@@ -2,4 +2,4 @@
 export HOME=/home/vagrant
 export WORKON_HOME=$HOME/.virtualenvs
 source /usr/local/bin/virtualenvwrapper.sh
-mkvirtualenv -a /home/vagrant/canvas_admin_tools -r /home/vagrant/canvas_admin_tools/canvas_admin_tools/requirements/local.txt canvas_admin_tools
+mkvirtualenv -a /home/vagrant/canvas_course_admin_tools -r /home/vagrant/canvas_course_admin_tools/canvas_course_admin_tools/requirements/local.txt canvas_course_admin_tools
